rename: common 폴더 생성
